Fix import path in Probe.ts
